move lru-cache to nextjs sdk
