+++ packages/sdks/overrides/rsc/src/helpers/preview-lru-cache/get.ts
@@ -1,0 +1,13 @@
+import { getIdFromSearchParams } from './helpers';
+import { init } from './init';
+import type { GlobalWCache } from './types';
+
+export function getPreviewContent(searchParams: URLSearchParams) {
+  init();
+
+  const id = getIdFromSearchParams(searchParams);
+
+  return typeof id === 'string'
+    ? (globalThis as GlobalWCache)._BUILDER_PREVIEW_LRU_CACHE.get(id)
+    : undefined;
+}
