+++ packages/sdks/overrides/rsc/src/helpers/preview-lru-cache/set.ts
@@ -1,4 +1,9 @@
+'use server';
+
 import type { BuilderContent } from '../../types/builder-content';
+import { init } from './init';
+import type { GlobalWCache } from './types';
+
 export async function postPreviewContent({
   key,
   value,
@@ -6,5 +11,9 @@ export async function postPreviewContent({
   key: string;
   value: BuilderContent;
 }) {
+  init();
+
+  (globalThis as GlobalWCache)._BUILDER_PREVIEW_LRU_CACHE.set(key, value);
+
   return { [key]: value };
 }
